+++ app/Http/Controllers/Admin/QuizController.php
@@ -3,14 +3,16 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Quiz;
-use App\Models\QuizAction;
 use App\Models\QuizTag;
+use App\Models\Role;
 use App\Models\Tag;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\Response;
 
 class QuizController extends AdminController
@@ -19,10 +21,14 @@ class QuizController extends AdminController
     protected string $sectionName = 'quizzes';
 
     /**
-     * @return Application|Factory|View
+     * @return Application|Factory|View|JsonResponse
      */
     public function index()
     {
+        if (!$this->checkAccess('quizzes', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/quizzes/list', [
             'sectionName' => $this->sectionName,
             'quizzes' => Quiz::query()->paginate(self::ITEM_ON_PAGE)
@@ -32,10 +38,14 @@ class QuizController extends AdminController
     /**
      * @param Request $request
      * @param $id
-     * @return Application|Factory|View
+     * @return Application|Factory|View|JsonResponse
      */
     public function edit(Request $request, $id)
     {
+        if (!$this->checkAccess('quizzes', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:1000',
@@ -89,10 +99,14 @@ class QuizController extends AdminController
 
     /**
      * @param Request $request
-     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
      */
     public function create(Request $request)
     {
+        if (!$this->checkAccess('quizzes', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:1000',
@@ -126,10 +140,14 @@ class QuizController extends AdminController
 
     /**
      * @param $id
-     * @return Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @return Application|JsonResponse|RedirectResponse|Redirector
      */
     public function delete($id)
     {
+        if (!$this->checkAccess('quizzes', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         Quiz::findOrFail($id)->delete();
 
         return redirect('/admin/quizzes/list');
