+++ app/Http/Controllers/Admin/StudyController.php
@@ -5,31 +5,52 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Book;
 use App\Models\Material;
+use App\Models\Role;
 use App\Models\Site;
 use App\Models\Topic;
 use App\Models\Video;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 
 class StudyController extends AdminController
 {
     protected string $sectionName = 'study';
 
     /**
-     * @return Application|Factory|View
+     * @return Application|Factory|View|JsonResponse
      */
     public function index()
     {
+        if (
+            !$this->checkAccess('study_videos', Role::showAccessType)
+            && !$this->checkAccess('study_books', Role::showAccessType)
+            && !$this->checkAccess('study_materials', Role::showAccessType)
+            && !$this->checkAccess('study_sites', Role::showAccessType)
+        ) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/study/list', [
             'sectionName' => $this->sectionName,
             'topics' => Topic::query()->paginate(self::ITEM_ON_PAGE),
         ]);
     }
 
+    /**
+     * @param $topicId
+     * @return Application|Factory|View|JsonResponse
+     */
     public function videos($topicId)
     {
+        if (!$this->checkAccess('study_videos', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/study/videos/videos', [
             'sectionName' => $this->sectionName,
             'videos' => Video::query()->where('topic_id', $topicId)->paginate(self::ITEM_ON_PAGE),
@@ -37,8 +58,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $topicId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
+     */
     public function createVideo($topicId, Request $request)
     {
+        if (!$this->checkAccess('study_videos', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $videoModel = new Video();
 
         if ($request->isMethod('post')) {
@@ -74,8 +104,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $videoId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse
+     */
     public function editVideo($videoId, Request $request)
     {
+        if (!$this->checkAccess('study_videos', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $videoModel = Video::findORFail($videoId);
 
         if ($request->isMethod('post')) {
@@ -109,8 +148,16 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $videoId
+     * @return Application|JsonResponse|RedirectResponse|Redirector
+     */
     public function deleteVideo($videoId)
     {
+        if (!$this->checkAccess('study_videos', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $videoModel = Video::findORFail($videoId);
         $topicId = $videoModel->topic_id;
         $videoModel->delete();
@@ -118,8 +165,16 @@ class StudyController extends AdminController
         return redirect('/admin/study/list/' . $topicId . '/videos');
     }
 
+    /**
+     * @param $topicId
+     * @return Application|Factory|View|JsonResponse
+     */
     public function books($topicId)
     {
+        if (!$this->checkAccess('study_books', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/study/books/books', [
             'sectionName' => $this->sectionName,
             'books' => Book::query()->where('topic_id', $topicId)->paginate(self::ITEM_ON_PAGE),
@@ -127,8 +182,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $topicId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
+     */
     public function createBook($topicId, Request $request)
     {
+        if (!$this->checkAccess('study_books', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $bookModel = new Book();
 
         if ($request->isMethod('post')) {
@@ -172,8 +236,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $bookId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse
+     */
     public function editBook($bookId, Request $request)
     {
+        if (!$this->checkAccess('study_books', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $bookModel = Book::findORFail($bookId);
 
         if ($request->isMethod('post')) {
@@ -214,8 +287,16 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $bookId
+     * @return Application|JsonResponse|RedirectResponse|Redirector
+     */
     public function deleteBook($bookId)
     {
+        if (!$this->checkAccess('study_books', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $bookModel = Book::findORFail($bookId);
         $topicId = $bookModel->topic_id;
         $bookModel->delete();
@@ -223,8 +304,16 @@ class StudyController extends AdminController
         return redirect('/admin/study/list/' . $topicId . '/books');
     }
 
+    /**
+     * @param $topicId
+     * @return Application|Factory|View|JsonResponse
+     */
     public function materials($topicId)
     {
+        if (!$this->checkAccess('study_materials', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/study/materials/materials', [
             'sectionName' => $this->sectionName,
             'materials' => Material::query()->where('topic_id', $topicId)->paginate(self::ITEM_ON_PAGE),
@@ -232,8 +321,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $topicId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
+     */
     public function createMaterial($topicId, Request $request)
     {
+        if (!$this->checkAccess('study_materials', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $materialModel = new Material();
 
         if ($request->isMethod('post')) {
@@ -273,8 +371,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $bookId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse
+     */
     public function editMaterial($bookId, Request $request)
     {
+        if (!$this->checkAccess('study_materials', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $materialModel = Material::findORFail($bookId);
 
         if ($request->isMethod('post')) {
@@ -311,8 +418,16 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $materialId
+     * @return Application|JsonResponse|RedirectResponse|Redirector
+     */
     public function deleteMaterial($materialId)
     {
+        if (!$this->checkAccess('study_materials', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $materialModel = Material::findORFail($materialId);
         $topicId = $materialModel->topic_id;
         $materialModel->delete();
@@ -320,8 +435,16 @@ class StudyController extends AdminController
         return redirect('/admin/study/list/' . $topicId . '/materials');
     }
 
+    /**
+     * @param $topicId
+     * @return Application|Factory|View|JsonResponse
+     */
     public function sites($topicId)
     {
+        if (!$this->checkAccess('study_sites', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/study/sites/sites', [
             'sectionName' => $this->sectionName,
             'sites' => Site::query()->where('topic_id', $topicId)->paginate(self::ITEM_ON_PAGE),
@@ -329,8 +452,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $topicId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
+     */
     public function createSite($topicId, Request $request)
     {
+        if (!$this->checkAccess('study_sites', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $siteModel = new Site();
 
         if ($request->isMethod('post')) {
@@ -366,8 +498,17 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $videoId
+     * @param Request $request
+     * @return Application|Factory|View|JsonResponse
+     */
     public function editSite($videoId, Request $request)
     {
+        if (!$this->checkAccess('study_sites', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $siteModel = Site::findORFail($videoId);
 
         if ($request->isMethod('post')) {
@@ -401,8 +542,16 @@ class StudyController extends AdminController
         ]);
     }
 
+    /**
+     * @param $siteId
+     * @return Application|JsonResponse|RedirectResponse|Redirector
+     */
     public function deleteSite($siteId)
     {
+        if (!$this->checkAccess('study_sites', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         $siteModel = Site::findORFail($siteId);
         $topicId = $siteModel->topic_id;
         $siteModel->delete();
