+++ app/Http/Controllers/Admin/TagController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Models\Role;
 use App\Models\Tag;
 use Illuminate\Http\Request;
 
@@ -11,6 +12,10 @@ class TagController extends AdminController
 
     public function index()
     {
+        if (!$this->checkAccess('tags', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/tags/list', [
             'sectionName' => $this->sectionName,
             'tags' => Tag::query()->paginate(self::ITEM_ON_PAGE)
@@ -19,6 +24,10 @@ class TagController extends AdminController
 
     public function create(Request $request)
     {
+        if (!$this->checkAccess('tags', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:100',
@@ -34,6 +43,10 @@ class TagController extends AdminController
 
     public function edit(int $id, Request $request)
     {
+        if (!$this->checkAccess('tags', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:100',
@@ -58,6 +71,10 @@ class TagController extends AdminController
 
     public function delete($id)
     {
+        if (!$this->checkAccess('tags', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         Tag::findOrFail($id)->delete();
 
         return redirect('/admin/tags/list');
