+++ app/Http/Controllers/Admin/TopicController.php
@@ -3,19 +3,30 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Role;
 use App\Models\Tag;
 use App\Models\Topic;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 
 class TopicController extends AdminController
 {
     protected string $sectionName = 'topics';
 
     /**
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @return Application|Factory|View|JsonResponse
      */
     public function index()
     {
+        if (!$this->checkAccess('topics', Role::showAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         return view('admin/topics/list', [
             'sectionName' => $this->sectionName,
             'topics' => Topic::query()->paginate(self::ITEM_ON_PAGE)
@@ -24,10 +35,14 @@ class TopicController extends AdminController
 
     /**
      * @param Request $request
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @return Application|JsonResponse|RedirectResponse|Redirector
      */
     public function create(Request $request)
     {
+        if (!$this->checkAccess('topics', Role::createAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:100',
@@ -44,10 +59,14 @@ class TopicController extends AdminController
     /**
      * @param int $id
      * @param Request $request
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @return Application|Factory|View|JsonResponse|RedirectResponse|Redirector
      */
     public function edit(int $id, Request $request)
     {
+        if (!$this->checkAccess('topics', Role::updateAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         if ($request->isMethod('post')) {
             $request->validate([
                 'name' => 'required|max:100',
@@ -72,10 +91,14 @@ class TopicController extends AdminController
 
     /**
      * @param $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @return Application|JsonResponse|RedirectResponse|Redirector
      */
     public function delete($id)
     {
+        if (!$this->checkAccess('topics', Role::deleteAccessType)) {
+            return response()->json(["error" => "Access denied for your role"], 403);
+        }
+
         Topic::findOrFail($id)->delete();
 
         return redirect('/admin/topics/list');
